app/requirements.txt
python-telegram-bot==13.11
requests
python-dotenv
ffmpeg-python
apng
pillow
signalstickers_client
asyncio
